# ExprNative interface - Allow easier use of custom types that can be represented by go builtin types