@antonmedv 2023-11-01
I understand your idea, clever solution. But penalty for hot path is way too big. Also, the API naming is little bit confusing. 

Lets try to find a solution without compromising on speed. 

I see you used  `testOne.Value()` in call benchmark. Lets try to replace it with another pattern:

```
getValue(testOne)
```

And getValue should be of type `func (any) any`, this way Expr can use fast call opcode. 

Use a expr.Patch to find all variable what implement the interface and patch with a function call.