@rrb3942 2023-11-02
Thank you @antonmedv, that is really good feedback and guidance. I wasn't very happy with the hot path change but wasn't aware of a better approach, so I appreciate it.

I'll work on switching this over to a solution using expr.Patch.