@rrb3942 2023-11-03
The problem that I'm running into with this approach is that type mismatches are not caught at compile time and only at runtime. It's very important for me that as many issues be caught at compile time.

Is there a way to do this that would also hint the compiler about the returned type so type issues could be caught?