@rrb3942 2023-11-04
Here is an example pattern that I came up with using interfaces:
https://go.dev/play/p/WpdiBXQHcdH

```
package main

import (
	"fmt"
	"reflect"

	"github.com/antonmedv/expr"
	"github.com/antonmedv/expr/ast"
)

// Generic fetch function for all types
type Valuer interface {
	Value() any
}

// Example interface if multiple types return this value type.
// Instead of iterating every type in the function declaration we use a more specific
// interface that embeds our generic interface and the underlying types implement both
type IntValue interface {
	Valuer
	IntValue() int
}

type FloatValue interface {
	Valuer
	FloatValue() float64
}

// Same as above but for strings
type StringValue interface {
	Valuer
	StringValue() string
}

type MyInt struct {
	Int int
}

func (v *MyInt) IntValue() int {
	return v.Int
}

func (v *MyInt) Value() any {
	return v.Int
}

type MyString struct {
	String string
}

func (v *MyString) StringValue() string {
	return v.String
}

func (v *MyString) Value() any {
	return v.String
}

// Only implements Valuer type must be explicitly defined on function
type MyFloat struct {
	Float float64
}

func (v *MyFloat) Value() any {
	return v.Float
}

type Env struct {
	Value any
}

var valuerType = reflect.TypeOf((*Valuer)(nil)).Elem()

type getValuePatcher struct{}

func (getValuePatcher) Visit(node *ast.Node) {
	id, ok := (*node).(*ast.IdentifierNode)
	if !ok {
		return
	}

	if id.Type().Implements(valuerType) {
		ast.Patch(node, &ast.CallNode{
			Callee:    &ast.IdentifierNode{Value: "getValue"},
			Arguments: []ast.Node{id},
		})
	}
}

func main() {
	code := `ValueOne + 1`

	env := make(map[string]any)
	env["ValueOne"] = &MyInt{1}
	env["ValueTwo"] = &MyString{"astring"}
	env["ValueThree"] = &MyFloat{1.2}

	program, err := expr.Compile(code,
		expr.Env(env),
		expr.Function("getValue", func(params ...any) (any, error) {
			// Use our generic fetch function to return the value
			return params[0].(Valuer).Value(), nil
		},
			// Now we can have it check for specific interface types as an input
			// and set the right return value type for checking
			new(func(IntValue) int),
			new(func(StringValue) string),
			// Or explicit types, make sure they implement the generic fetch
			new(func(*MyFloat) float64)),
		expr.Patch(getValuePatcher{}),
	)
	if err != nil {
		panic(err)
	}

	output, err := expr.Run(program, env)
	if err != nil {
		panic(err)
	}

	fmt.Println(output)
}

```

Still need to run some benchmarks and see how it compares.