@rrb3942 2023-11-10
The patching methods works very well for my use case and performance is good. I'm going to close this PR since it can be accomplished already via patching.

If you think it would be helpful to have a builtin interface type for others to use, or want to have some builtin patchers available by default I'd be happy to redo this using the patching method.

Thanks for all the help!