@rrb3942 2023-11-15
> It is possible! You can set type in the patcher. Or you can set type in expr.Function:
> 
> ```diff
>  expr.Function("getValue", func(params ...any) (any, error) {
>  	return params[0].(Value).Int, nil
>  },
> +new(func (Value) int)),
> ```

Do you have an example of how to set the type in the patcher instead of the expr.Function?