@antonmedv 2023-11-15
Sure, try:
```go
node.SetType()
```