@rrb3942 2023-11-15
> Sure, try:
> 
> ```go
> node.SetType()
> ```

That's what I was trying to use, but it doesn't seem to help catch type mismatches at compile time. I just assume I'm doing something wrong.

https://go.dev/play/p/hori9Xtni-k

Panic happens at `Run` and not `Compile`

I also tried using a type of `new(func (Value) int))`.