@antonmedv 2023-11-16
I see. This is definitely strange. I will expect what setting type in Patch will take effect here. 

Created a issue to track this bug.