@rrb3942 2023-11-18
If you wouldn't mind having a look at:

https://github.com/rrb3942/expr/commit/0572429028dec548824d162ef6ebd6c94d4af447

And give feedback if you think the overall approach looks good. Still working on adding more testing and documentation. Wasn't sure where to put it. Maybe just 'expr/patchers/value' instead of 'expr/builtin/patchers/value'? Or something else entirely.

Thanks!